Mod: cambiati offset da int a float.

+++ RM250714_RobotPanini/src/RM250714/Classes/FR20/RobotManager.cs
@@ -3773,37 +3773,37 @@ namespace RM.src.RM250714
 
             #region Offset pick
 
-            int offsetAvvicinamentoPick = 5; // Offset applicato al punto di avvicinamento pick
-            int zOffsetAvvicinamentoPick = 10; // Offset su asse Z in cui mi abbasso leggermente prima di andare in pick
-            int zOffsetPostPick = 20; // Offset applicato dopo chiusura pinza (mantengo questo offset anche durante movimento di allontanamento dal pick)
-            int zOffsetAllontanamentoPick = 20; // Offset applicato per punto di allontanamento pick
-            int offsetAllontamentoIntPick = 250; // Offset necessario per avere un punto intermedio di allontamento post pick
-            int offsetAllontamentoPick = 770; // Offset  di allontamento post pick
-            int zOffsetPick = 3; // Offset del punto di pick su asse z
-            int offsetAllontamentoPreSlittaIndietro = 0; // Offset utilizzato per sapere quanto dopo aver eseguito il pick inviare il comando di slitta indietro
-            int yOffsetPick = 5; // Offset del punto di pick su asse y
-            int offsetPreAvvicinamentoPick = 130; // Offset di avvicinamento al punto di avvicinamento pick
-            int rxRotationPick = 3; // Gradi di rotazione su asse rx dopo aver eseguito il pick
+            float offsetAvvicinamentoPick = 5; // Offset applicato al punto di avvicinamento pick
+            float zOffsetAvvicinamentoPick = 10; // Offset su asse Z in cui mi abbasso leggermente prima di andare in pick
+            float zOffsetPostPick = 20; // Offset applicato dopo chiusura pinza (mantengo questo offset anche durante movimento di allontanamento dal pick)
+            float zOffsetAllontanamentoPick = 20; // Offset applicato per punto di allontanamento pick
+            float offsetAllontamentoIntPick = 250; // Offset necessario per avere un punto intermedio di allontamento post pick
+            float offsetAllontamentoPick = 770; // Offset  di allontamento post pick
+            float zOffsetPick = 3; // Offset del punto di pick su asse z
+            float offsetAllontamentoPreSlittaIndietro = 0; // Offset utilizzato per sapere quanto dopo aver eseguito il pick inviare il comando di slitta indietro
+            float yOffsetPick = 5; // Offset del punto di pick su asse y
+            float offsetPreAvvicinamentoPick = 130; // Offset di avvicinamento al punto di avvicinamento pick
+            float rxRotationPick = 3; // Gradi di rotazione su asse rx dopo aver eseguito il pick
 
             #endregion
 
             #region Offset place
 
-            int offsetAvvicinamentoPlace = 770; // Offset per eseguire punto di avvicinamento place
-            int zOffsetAvvicinamentoPlace = 40; // Offset su asse Z in cui mi alzo leggermente prima di andare in place
-            int zOffsetPostPlace = 5; // Offset su asse Z in cui mi abbasso leggermente dopo essere andato in place
-            int offsetAllontamentoPostPlace = 130; // Offset di allontanamento dal carrello dopo aver eseguito il place
-            int offsetAllontamentoPreSlittaAvanti = 550; // Offset utilizzato per sapere quanto prima di raggiungere il place inviare il comando di slitta avanti
-            int zOffsetPlace = 10; // Offset del punto di place su asse z
-            int yOffsetPlace = 40; // Offset del punto di place su asse y
-            int rxOffsetPrePlace = 2; // Offset di rotazione su asse x applicato al punto di avvicinamento place
+            float offsetAvvicinamentoPlace = 770; // Offset per eseguire punto di avvicinamento place
+            float zOffsetAvvicinamentoPlace = 40; // Offset su asse Z in cui mi alzo leggermente prima di andare in place
+            float zOffsetPostPlace = 5; // Offset su asse Z in cui mi abbasso leggermente dopo essere andato in place
+            float offsetAllontamentoPostPlace = 130; // Offset di allontanamento dal carrello dopo aver eseguito il place
+            float offsetAllontamentoPreSlittaAvanti = 550; // Offset utilizzato per sapere quanto prima di raggiungere il place inviare il comando di slitta avanti
+            float zOffsetPlace = 10; // Offset del punto di place su asse z
+            float yOffsetPlace = 40; // Offset del punto di place su asse y
+            float rxOffsetPrePlace = 2; // Offset di rotazione su asse x applicato al punto di avvicinamento place
 
             #endregion
 
             #region Offset beor
 
-            int offsetAvvicinamentoBeor = 600; // Offset di avvicinamento al punto beor
-            int offsetAllontamentoBeor = 700; // Offset di allontanamento dal punto beor
+            float offsetAvvicinamentoBeor = 600; // Offset di avvicinamento al punto beor
+            float offsetAllontamentoBeor = 700; // Offset di allontanamento dal punto beor
 
             #endregion
